requirements-app.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
unsloth[cu121-torch251]
langgraph==0.2.24
langchain==0.3.2
google-cloud-speech==2.26.0
aiohttp==3.10.5
websockets==12.0
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
scipy==1.14.1
onnxruntime-gpu==1.17.1
pyyaml==6.0.2
python-dotenv==1.0.1
pydub==0.25.1
structlog==24.1.0
uvloop==0.20.0
transformers==4.36.2
huggingface_hub==0.22.2
datasets==2.14.7
pyarrow==12.0.1

# Asterisk REST Interface

# Observability
opentelemetry-distro==0.46b0
opentelemetry-instrumentation-aiohttp-client==0.46b0